requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
feedparser==6.0.10
psycopg2-binary==2.9.9
python-dotenv==1.0.0